requirements.txt
asyncio-paho
confuse
dnspython
icmplib~=3.0
python-box~=6.0
python-slugify
types-python-slugify